fix(ui): align chart icon and page-size chevron
