+++ chat2db-community-client/src/pages/main/CommunityMainPage.tsx
@@ -1,8 +1,9 @@
 import { Confetti, IconButton, IconfontSvg } from '@chat2db/ui';
 import { Tooltip, type InputRef } from 'antd';
-import { createLucideIcon, Layers, MessagesSquare } from 'lucide-react';
+import { Layers, MessagesSquare } from 'lucide-react';
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 
+import ChartNoAxesCombined from '@/components/LucideIcons/ChartNoAxesCombined';
 import i18n from '@/i18n';
 import { INavItem } from '@/typings/main';
 import feedback from '@/utils/feedback';
@@ -33,16 +34,6 @@ import { useChatStore } from '@/store/chat';
 import { useWorkspaceStore } from '@/store/workspace';
 import { isDesktop, isHashHistoryEnv } from '@/utils/env';
 import { checkIsSharePage } from '@/utils/url';
-
-// Backport the official Lucide icon; the direct 0.356 dependency predates this export.
-const ChartNoAxesCombined = createLucideIcon('ChartNoAxesCombined', [
-  ['path', { d: 'M12 16v5', key: 'zza2cw' }],
-  ['path', { d: 'M16 14v7', key: '1g90b9' }],
-  ['path', { d: 'M20 10v11', key: '1iqoj0' }],
-  ['path', { d: 'm22 3-8.646 8.646a.5.5 0 0 1-.708 0L9.354 8.354a.5.5 0 0 0-.707 0L2 15', key: '1fw8x9' }],
-  ['path', { d: 'M4 18v3', key: '1yp0dc' }],
-  ['path', { d: 'M8 14v7', key: 'n3cwzv' }],
-]);
 
 function CommunityMainPage() {
   const [navConfig, setNavConfig] = useState<INavItem[]>([]);
